requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
python-dotenv==1.1.0
openpyxl==3.1.5
google-genai==1.14.0
